feat(api): add validation for camera endpoints
Excluding the auth key field in the Camera Pydantic models. I'm not sure
how to handle authentication just yet.
Also renamed the 'camera_ids' and 'camera_id' parameters to 'id'.
'camera_ids' is a query parameter of a list type, which is used like the
following: '<domain-name>/api/v0/cameras?camera_ids=1&camera_ids=2',
which is why 'id' is a better name in my opinion.