feat(security): add password validation and argon2 hashing
Passwords must follow the following criteria:
- At least 8 characters long
- At least 1 lower and upper case character
- At least 1 number
- At least 1 special character (e.g. !?@)

Argon2 was used to hash the passwords as it is known to be difficult to
brute force it (as of 2025).
The password hash stored in the database will also contain the salt used
to generate the hash (so it can be verified) in the format:
'{salt}:{hashed-password}'.